+++ tests/testAutopas/ParticleIteratorTest.h
@@ -27,7 +27,7 @@ protected:
 	// needs to be protected, because the test fixtures generate a derived class for each unit test.
 
 	std::vector<autopas::MoleculeLJ> _vecOfMolecules;
-	int _currentIndex;
+	unsigned long _currentIndex;
 };
 
 #endif /* TESTS_TESTAUTOPAS_PARTICLEITERATORTEST_H_ */
